+++ src/api/ActivitiesApiDev.ts
@@ -12,7 +12,7 @@ export default class ActivitiesApiDev implements IActivityApi {
 			WeekOf: DateUtilities.getStartOfWeek().toISOString(), 
 			Branch: 'OZI',
 			ActionTaken: 'Lorem\n ipsum\n dolor sit amet, consectetur adipiscing elit.\n Morbi euismod lacus ac sagittis mollis. Nulla ut quam sed nisl pulvinar cursus sit amet eget lacus. Suspendisse rutrum pulvinar tortor ut vehicula. Nunc non arcu imperdiet, semper urna at, facilisis lectus. Phasellus risus magna, dignissim vel consequat ac, tincidunt in lacus. Aliquam euismod fringilla mauris, ac bibendum quam pulvinar vitae. Donec iaculis accumsan mi sed tincidunt. Proin accumsan, massa vitae malesuada porta, mauris purus facilisis sem, vel laoreet magna urna eget nulla. Phasellus convallis ipsum a convallis tincidunt.\n\nNam in leo velit. Mauris at ullamcorper leo. In tortor ligula, efficitur et diam sit amet, tincidunt finibus ligula. Aliquam finibus egestas justo ut posuere. Vestibulum pharetra, tellus et finibus pellentesque, dui leo consectetur augue, sit amet pharetra nisl velit sed sapien. Quisque non nunc turpis. Donec eu erat mauris. In et tincidunt enim. Donec luctus eu lectus sed scelerisque. Nulla iaculis ultricies lectus, nec eleifend ipsum auctor a. Quisque sed massa eros.', 
-			IsBigRock: false, 
+			IsMarEntry: false, 
 			IsHistoryEntry: true,
 			OPRs: { results: [{ Id: '1', Title: 'Robert Porterfield' }, { Id: '2', Title: 'Jeremy Clark' }] },
 			__metadata: {etag: '"1"'}
@@ -23,7 +23,7 @@ export default class ActivitiesApiDev implements IActivityApi {
 			WeekOf: DateUtilities.getStartOfWeek().toISOString(), 
 			Branch: 'OZI',
 			ActionTaken: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi euismod lacus ac sagittis mollis. Nulla ut quam sed nisl pulvinar cursus sit amet eget lacus. Suspendisse rutrum pulvinar tortor ut vehicula. Nunc non arcu imperdiet, semper urna at, facilisis lectus. Phasellus risus magna, dignissim vel consequat ac, tincidunt in lacus. Aliquam euismod fringilla mauris, ac bibendum quam pulvinar vitae. Donec iaculis accumsan mi sed tincidunt. Proin accumsan, massa vitae malesuada porta, mauris purus facilisis sem, vel laoreet magna urna eget nulla. Phasellus convallis ipsum a convallis tincidunt.\n\nNam in leo velit. Mauris at ullamcorper leo. In tortor ligula, efficitur et diam sit amet, tincidunt finibus ligula. Aliquam finibus egestas justo ut posuere. Vestibulum pharetra, tellus et finibus pellentesque, dui leo consectetur augue, sit amet pharetra nisl velit sed sapien. Quisque non nunc turpis. Donec eu erat mauris. In et tincidunt enim. Donec luctus eu lectus sed scelerisque. Nulla iaculis ultricies lectus, nec eleifend ipsum auctor a. Quisque sed massa eros.', 
-			IsBigRock: false, 
+			IsMarEntry: false, 
 			IsHistoryEntry: false,
 			OPRs: { results: [{ Id: '1', Title: 'Robert Porterfield' }] },
 			__metadata: { etag: '"1"' }
@@ -34,7 +34,7 @@ export default class ActivitiesApiDev implements IActivityApi {
 			WeekOf: DateUtilities.getStartOfWeek().toISOString(), 
 			Branch: 'OZIC',
 			ActionTaken: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi euismod lacus ac sagittis mollis. Nulla ut quam sed nisl pulvinar cursus sit amet eget lacus. Suspendisse rutrum pulvinar tortor ut vehicula. Nunc non arcu imperdiet, semper urna at, facilisis lectus. Phasellus risus magna, dignissim vel consequat ac, tincidunt in lacus. Aliquam euismod fringilla mauris, ac bibendum quam pulvinar vitae. Donec iaculis accumsan mi sed tincidunt. Proin accumsan, massa vitae malesuada porta, mauris purus facilisis sem, vel laoreet magna urna eget nulla. Phasellus convallis ipsum a convallis tincidunt.\n\nNam in leo velit. Mauris at ullamcorper leo. In tortor ligula, efficitur et diam sit amet, tincidunt finibus ligula. Aliquam finibus egestas justo ut posuere. Vestibulum pharetra, tellus et finibus pellentesque, dui leo consectetur augue, sit amet pharetra nisl velit sed sapien. Quisque non nunc turpis. Donec eu erat mauris. In et tincidunt enim. Donec luctus eu lectus sed scelerisque. Nulla iaculis ultricies lectus, nec eleifend ipsum auctor a. Quisque sed massa eros.', 
-			IsBigRock: true, 
+			IsMarEntry: true, 
 			IsHistoryEntry: true,
 			OPRs: { results: [{ Id: '2', Title: 'Jeremy Clark' }] },
 			__metadata: { etag: '"1"' }
@@ -60,9 +60,9 @@ export default class ActivitiesApiDev implements IActivityApi {
 		return this.activities.filter(activity => query ? activity.ActionTaken.includes(query) : true);
 	}
 
-	async fetchBigRocksByDates(startDate: Moment, endDate: Moment, userId: number) {
+	async fetchMarEntriesByDates(startDate: Moment, endDate: Moment, userId: number) {
 		await this.sleep();
-		return this.activities.filter(activity => activity.IsBigRock);
+		return this.activities.filter(activity => activity.IsMarEntry);
 	}
 
 	async fetchHistoryEntriesByDates(startDate: Moment, endDate: Moment, userId: number) {